changes for renaming the controller files
